--- lite/valor_lite/__init__.py
@@ -1,3 +1,0 @@
-from .detection import DetectionManager
-
-__all__ = ["DetectionManager"]
